feat: implement student subjects dashboard page and server-side data fetching
