+++ README.md
@@ -1,31 +1,56 @@
-﻿# Student Records Frontend
+# Student Records Frontend
 
-Frontend application for the Student Records platform.
+Frontend application for the Student Records school management platform.
 
 ## Stack
 
-- Next.js 16 (App Router)
-- TypeScript
-- Tailwind CSS v4
-- shadcn/ui component system
+- **Framework** — Next.js 16 (App Router)
+- **Language** — TypeScript (strict mode)
+- **Styling** — Tailwind CSS v4
+- **Components** — shadcn/ui (Base UI primitives)
+- **Auth & Database** — Supabase (Google OAuth, Row-Level Security)
+- **Testing** — Vitest + V8 coverage
+
+## Project Structure
+
+```
+src/
+├── app/                  # Next.js App Router pages & layouts
+│   ├── (dashboard)/      # Authenticated dashboard routes
+│   │   ├── admin/        # Admin workspace (allowlist, academic setup, users)
+│   │   ├── teacher/      # Teacher workspace
+│   │   ├── student/      # Student workspace
+│   │   └── subjects/     # Student subject view
+│   ├── auth/             # OAuth callback handler
+│   └── login/            # Login page
+├── components/ui/        # shadcn/ui component library
+├── features/             # Feature modules (domain logic)
+│   ├── academic-setup/   # Academic year, terms, subjects, enrollment
+│   ├── access-control/   # Role-based route permissions
+│   ├── allowlist/        # Allowlist management (single + bulk)
+│   ├── auth/             # Authentication helpers & guards
+│   ├── subjects/         # Student subject view logic
+│   └── user-management/  # User CRUD & filtering
+└── lib/                  # Shared utilities
+    ├── supabase/         # Supabase client/server/config
+    └── utils.ts          # Tailwind class merge helper
+```
 
 ## Getting Started
 
-Run the development server:
-
 ```bash
-npm run dev
+cp .env.example .env     # Fill in your Supabase credentials
+npm install
+npm run dev              # http://localhost:3000
 ```
-
-Open [http://localhost:3000](http://localhost:3000) with your browser to see the result.
-
-You can start editing the page in `src/app/page.tsx`.
 
 ## Quality Checks
 
 ```bash
-npm run lint
-npm run build
+npm run lint             # ESLint (flat config)
+npm run typecheck        # TypeScript strict
+npm run test             # Vitest
+npm run build            # Production build
 ```
 
 ## Add More shadcn Components
@@ -33,10 +58,3 @@ npm run build
 ```bash
 npx shadcn@latest add button card table dialog form select
 ```
-
-## Learn More
-
-To learn more about Next.js, take a look at the following resources:
-
-- [Next.js Documentation](https://nextjs.org/docs) - learn about Next.js features and API.
-- [Learn Next.js](https://nextjs.org/learn) - an interactive Next.js tutorial.
