+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "web",
+  "name": "student-records",
   "version": "0.1.0",
   "private": true,
   "scripts": {
